refactor(web): centralize clipboard helper and remove duplicate implementations
