+++ apps/web/src/components/workspace/file-preview-panel.tsx
@@ -4,8 +4,8 @@ import { useCallback, useState } from 'react'
 import { Check, ClipboardText, Eye, PencilSimple, X } from '@phosphor-icons/react'
 
 import { Button } from '@/components/ui/button'
-
-import { MarkdownPreview } from './markdown-preview'
+import { MarkdownPreview } from '@/components/workspace/markdown-preview'
+import { copyTextToClipboard } from '@/lib/clipboard'
 
 type FilePreviewPanelProps = {
   path: string
@@ -13,35 +13,6 @@ type FilePreviewPanelProps = {
   isLoading?: boolean
   onClose: () => void
   onEdit: () => void
-}
-
-async function copyTextToClipboard(text: string): Promise<boolean> {
-  if (typeof navigator !== 'undefined' && navigator.clipboard?.writeText) {
-    try {
-      await navigator.clipboard.writeText(text)
-      return true
-    } catch {
-      // Fall through to legacy fallback below.
-    }
-  }
-  if (typeof document === 'undefined') return false
-  const textarea = document.createElement('textarea')
-  textarea.value = text
-  textarea.setAttribute('readonly', '')
-  textarea.style.position = 'fixed'
-  textarea.style.top = '-9999px'
-  textarea.style.opacity = '0'
-  document.body.appendChild(textarea)
-  textarea.focus()
-  textarea.select()
-  let ok = false
-  try {
-    ok = document.execCommand('copy')
-  } catch {
-    ok = false
-  }
-  document.body.removeChild(textarea)
-  return ok
 }
 
 export function FilePreviewPanel({
